requirements.txt
# Backend API
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Machine Learning
tensorflow>=2.16.0
numpy>=1.24.3
pillow>=10.1.0

# Análisis y visualización (para notebooks)
pandas>=2.1.3
matplotlib>=3.8.2

# Utilidades
pathlib2==2.3.7; python_version < '3.4'

